LayoutTests/http/tests/security/img-with-failed-cors-check-fails-to-load-expected.txt

CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
ALERT: PASS: The error event was called.
This test passes if the image below does not load. 
